backend/services/auth_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.5.0
python-multipart==0.0.6
redis==5.0.1
minio==7.2.0
cryptography==41.0.8
httpx==0.25.2
python-keycloak==3.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
